back-end/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.1.1
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
slowapi==0.1.9
bcrypt==5.0.0
yagmail==0.15.293
pandas==2.3.3
numpy==1.26.4
openpyxl==3.1.5
openai==2.6.1
transformers==4.57.1
pillow==12.0.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.185.0
langchain-openai==1.0.1
langchain==1.0.2
httpx==0.28.1
requests==2.32.3
tiktoken==0.12.0
torch==2.9.0
chromadb==1.2.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
alembic==1.17.0
codecarbon==3.0.8